+++ src/test/java/io/github/thebusybiscuit/slimefun4/testing/tests/services/TestUpdaterService.java
@@ -5,6 +5,7 @@ import java.io.File;
 import org.junit.jupiter.api.AfterAll;
 import org.junit.jupiter.api.Assertions;
 import org.junit.jupiter.api.BeforeAll;
+import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 
 import be.seeseemelk.mockbukkit.MockBukkit;
@@ -12,7 +13,7 @@ import io.github.thebusybiscuit.slimefun4.api.SlimefunBranch;
 import io.github.thebusybiscuit.slimefun4.core.services.UpdaterService;
 import io.github.thebusybiscuit.slimefun4.implementation.SlimefunPlugin;
 
-public class TestUpdaterService {
+class TestUpdaterService {
 
     private static SlimefunPlugin plugin;
 
@@ -30,7 +31,8 @@ public class TestUpdaterService {
     }
 
     @Test
-    public void testDevelopmentBuilds() {
+    @DisplayName("Test if the development branch is recognized correctly")
+    void testDevelopmentBuilds() {
         UpdaterService service = new UpdaterService(plugin, "DEV - 131 (git 123456)", file);
         Assertions.assertEquals(SlimefunBranch.DEVELOPMENT, service.getBranch());
         Assertions.assertTrue(service.getBranch().isOfficial());
@@ -39,7 +41,8 @@ public class TestUpdaterService {
     }
 
     @Test
-    public void testStableBuilds() {
+    @DisplayName("Test if the stable branch is recognized correctly")
+    void testStableBuilds() {
         UpdaterService service = new UpdaterService(plugin, "RC - 6 (git 123456)", file);
         Assertions.assertEquals(SlimefunBranch.STABLE, service.getBranch());
         Assertions.assertTrue(service.getBranch().isOfficial());
@@ -48,7 +51,8 @@ public class TestUpdaterService {
     }
 
     @Test
-    public void testUnofficialBuilds() {
+    @DisplayName("Test if an unofficial build is recognized correctly")
+    void testUnofficialBuilds() {
         UpdaterService service = new UpdaterService(plugin, "4.20 UNOFFICIAL", file);
         Assertions.assertEquals(SlimefunBranch.UNOFFICIAL, service.getBranch());
         Assertions.assertFalse(service.getBranch().isOfficial());
@@ -56,7 +60,8 @@ public class TestUpdaterService {
     }
 
     @Test
-    public void testUnknownBuilds() {
+    @DisplayName("Test if unknown builds are caught")
+    void testUnknownBuilds() {
         UpdaterService service = new UpdaterService(plugin, "I am special", file);
         Assertions.assertEquals(SlimefunBranch.UNKNOWN, service.getBranch());
         Assertions.assertFalse(service.getBranch().isOfficial());